[parser::mesh] Factor all execution out of the parser module to the loading module
