+++ bve/src/load/mesh/cube.obj
