--- bve/src/parse/mesh/instructions/tests/meshes/mod.rs
@@ -1,1 +1,0 @@
-pub const CUBE_SOURCE: &str = include_str!("cube.obj");
